wrap js with IIFE; standardize controller names
This change seems to improve the viability of the grunt build.
Previously the grunt build did not function as the application tree
functioned. Specifically the evidence search did not apply selected
filters to your next search.

This might have had to do with wrapping functions in a function (IIFE)
for scoping purposes.